requirements.txt
keras
tensorflow
flask
flask_cors
gevent
python-dotenv
pillow
requests
gunicorn
